1.removed system clock initialization and dwt counter intiialization from main 2.create mcu driver file to add clock and dwt initialization code here
